refactor(FE-10): added useImportTrip hook to incapsulate import logic
- coding and decoding logic moved to shared layer
