+++ src/features/share-trip/index.ts
@@ -1,5 +1,7 @@
 export { usePopup } from './lib/hooks/usePopup';
-export { useShareCode } from './lib/hooks/useShareCode';
 
 export { SharePopup } from './ui/SharePopup';
 export { ImportPopup } from './ui/ImportPopup';
+
+export { ImportStatus } from './lib/types';
+export type { ImageRaw, StopRaw, TripRaw } from './lib/types';
