+++ src/features/share-trip/lib/index.ts
@@ -1,4 +1,9 @@
 export { usePopup } from './hooks/usePopup';
 export { useShareCode } from './hooks/useShareCode';
+export { useImportTrip } from './hooks/useImportTrip';
 
 export { generateShareCode } from './utils/generateShareCode';
+export { parseShareCode } from './utils/parseShareCode';
+
+export { ImportStatus } from './types';
+export type { ImageRaw, StopRaw, TripRaw } from './types';
