+++ src/shared/lib/codec/index.ts
@@ -1,0 +1,1 @@
+export { compressData, decompressData } from './codec';
