+++ src/shared/lib/index.ts
@@ -3,3 +3,5 @@ export * from './dropdown';
 export * from './suspense';
 export * from './dnd';
 export * from './date';
+
+export { compressData, decompressData } from './codec';
